requirements.txt
amqp==2.3.2
anyjson==0.3.3
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==2.0.6
django-ajax-selects==1.7.0
django-appconf==1.0.2
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-celery-beat==1.1.1
django-constance==2.2.0
django-crispy-forms==1.7.2
django-excel==0.0.10
django-filter==2.0.0
django-forms-ajaxified==0.2
django-guardian==1.4.9
django-import-export==1.0.1
django-jquery==3.1.0
django-libs==2.0.1
django-multiselectfield==0.1.8
django-oauth-toolkit==1.2.0
django-picklefield==1.0.0
django-search-listview==0.2.1
django-search-views==0.3.7
django-select2==6.0.3
django-sendgrid-v5==0.6.893
django-tabination==0.4.0
django-widget-tweaks==1.4.2
et-xmlfile==1.0.1
fabric==2.3.1
future==0.16.0
idna==2.7
invoke==1.1.1
jdcal==1.4
kombu==4.2.1
lml==0.0.1
numpy==1.15.0
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.4
pandas==0.23.3
paramiko==2.4.1
pipenv==2018.7.1
psycopg2==2.7.4
pyasn1==0.4.4
pycparser==2.18
pyexcel==0.5.8
pyexcel-io==0.5.7
pyexcel-webio==0.1.4
pyexcel-xlsx==0.5.6
PyNaCl==1.2.1
python-dateutil==2.7.3
python-decouple==3.1
python-http-client==3.1.0
pytz==2018.4
PyYAML==3.13
redis==2.10.6
requests==2.19.1
sendgrid==5.5.0
six==1.11.0
tablib==0.12.1
texttable==1.4.0
unicodecsv==0.14.1
urllib3==1.23
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
xlrd==1.1.0
xlwt==1.3.0
